Display improvements, fix broken link.
Escape all '+' in key1 + key2 lines

E.g., change key1 + key2 to key1 \+ key2. This prevents markdown from
seeing this a special character, which can happen when a line join puts
the plus at the beginning of line.

When showing a sequence, change > to &rarr;. E.g. change:
 `File` > `Save`
to
 `File` &rarr; `Save`

N.B.: spell check sees this as just "rarr". Maybe I should add this to
spellright.dict?

Joined lines that got longer.

Fix broken link to #moving-on-the-line-and-up-and-down-lines. Previous
commit fixed spelling of "and" but didn't change the link.